store can now add Products store can now edit Products store can now delete products
added new views for adding and editing/deleting products
added the new views to the router so the store can go there.

added function to the server for inserting to new product the database.
added function to the server for updating a product in the database.
added function to the server for deleting a product in the database.
added function to the server for get all product in the database.